Removing engine_version validation
The validation we had was only valid for PostgreSQL and as a result
prevented MySQL clusters from upgrading to the latest version of this
module because their version string is formatted different
(8.0.mysql_aurora.3.07.0 for example).

I think it would be pretty tough to validate this field. For your
consideration, consider the output of
`aws rds describe-db-engine-versions`. At the time of writing, running
this command gives 154 options for mysql and postgres alone. Given that
terragrunt apply would catch this for us, I don't think it's really
worth trying to validate against so many options.